move and rename tests
